+++ Bank Infrastructure/src/main/java/com/exalt_company/kata_bank/adapter/v1/resource/FundRequest.java
@@ -1,5 +1,6 @@
 package com.exalt_company.kata_bank.adapter.v1.resource;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record FundRequest(UUID fundOwnerId, UUID fundId, double amount) {}
+public record FundRequest(UUID fundOwnerId, UUID fundId, BigDecimal amount) {}
